+++ app.js
@@ -1,5 +1,4 @@
 const express = require("express");
-const mongoose = require("mongoose");
 const routes = require("./routes/routes");
 const dbconnect = require("./config/db");
